requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
psycopg2-binary==2.9.9
requests==2.31.0
markdown==3.5.1
